Refactor server.ts into modular architecture (1665→281 lines)
Extract 6 logical modules from the monolithic server.ts file:

Services:
- port.service.ts: Port availability utilities
- session-file-logger.service.ts: WebSocket debug logging

Middleware:
- security.middleware.ts: Security headers
- cors.middleware.ts: CORS configuration
- request-logger.middleware.ts: HTTP request logging

API Routers:
- health.router.ts: Health check endpoints
- mcp.router.ts: MCP tool execution

WebSocket Handlers:
- chat.handler.ts: Chat WebSocket (~650 lines)
- terminal.handler.ts: Terminal WebSocket (~340 lines)
