+++ src/backend/middleware/index.ts
@@ -1,0 +1,3 @@
+export { corsMiddleware } from './cors.middleware';
+export { requestLoggerMiddleware } from './request-logger.middleware';
+export { securityMiddleware } from './security.middleware';
