+++ src/backend/middleware/request-logger.middleware.ts
@@ -1,0 +1,25 @@
+import type { NextFunction, Request, Response } from 'express';
+import { createLogger } from '../services/index';
+
+const logger = createLogger('server');
+
+/**
+ * Request logging middleware.
+ * Logs HTTP requests with method, path, status, and duration.
+ * Skips logging for /health endpoints to reduce noise.
+ */
+export function requestLoggerMiddleware(req: Request, res: Response, next: NextFunction): void {
+  const start = Date.now();
+  res.on('finish', () => {
+    const duration = Date.now() - start;
+    if (req.path !== '/health' && !req.path.startsWith('/health/')) {
+      logger.debug('HTTP request', {
+        method: req.method,
+        path: req.path,
+        status: res.statusCode,
+        duration,
+      });
+    }
+  });
+  next();
+}
