+++ src/backend/routers/api/health.router.ts
@@ -1,0 +1,90 @@
+import { Router } from 'express';
+import { prisma } from '../../db';
+import { configService, createLogger, rateLimiter } from '../../services/index';
+
+const router = Router();
+const logger = createLogger('api:health');
+
+// ============================================================================
+// Health Check Routes
+// ============================================================================
+
+/**
+ * GET /health
+ * Basic health check - returns service status and metadata
+ */
+router.get('/', (_req, res) => {
+  res.json({
+    status: 'ok',
+    timestamp: new Date().toISOString(),
+    service: 'factoryfactory-backend',
+    version: process.env.npm_package_version || '0.1.0',
+    environment: configService.getEnvironment(),
+  });
+});
+
+/**
+ * GET /health/database
+ * Database connectivity health check
+ */
+router.get('/database', async (_req, res) => {
+  try {
+    await prisma.$queryRaw`SELECT 1`;
+    res.json({
+      status: 'ok',
+      timestamp: new Date().toISOString(),
+      database: 'connected',
+    });
+  } catch (error) {
+    logger.error('Database health check failed', error as Error);
+    res.status(503).json({
+      status: 'error',
+      timestamp: new Date().toISOString(),
+      database: 'disconnected',
+      error: error instanceof Error ? error.message : 'Unknown error',
+    });
+  }
+});
+
+/**
+ * GET /health/all
+ * Comprehensive health check - database + rate limiter status
+ */
+router.get('/all', async (_req, res) => {
+  const checks: Record<string, { status: string; details?: unknown }> = {};
+
+  try {
+    await prisma.$queryRaw`SELECT 1`;
+    checks.database = { status: 'ok' };
+  } catch (error) {
+    checks.database = {
+      status: 'error',
+      details: error instanceof Error ? error.message : 'Unknown error',
+    };
+  }
+
+  const apiUsage = rateLimiter.getApiUsageStats();
+  checks.rateLimiter = {
+    status: apiUsage.isRateLimited ? 'degraded' : 'ok',
+    details: {
+      requestsLastMinute: apiUsage.requestsLastMinute,
+      isRateLimited: apiUsage.isRateLimited,
+    },
+  };
+
+  const statuses = Object.values(checks).map((c) => c.status);
+  let overallStatus = 'ok';
+  if (statuses.includes('error')) {
+    overallStatus = 'error';
+  } else if (statuses.includes('degraded')) {
+    overallStatus = 'degraded';
+  }
+
+  res.status(overallStatus === 'error' ? 503 : 200).json({
+    status: overallStatus,
+    timestamp: new Date().toISOString(),
+    checks,
+  });
+});
+
+export { router as healthRouter };
